+++ gen/com/intellij/idea/plugin/hybris/impex/ImpexParser.java
@@ -1580,6 +1580,7 @@ public class ImpexParser implements PsiParser, LightPsiParser {
   //     | FIELD_VALUE_JAR_PREFIX
   //     | FIELD_VALUE_FILE_PREFIX
   //     | FIELD_VALUE_ZIP_PREFIX
+  //     | FIELD_VALUE_HTTP_PREFIX
   //     | FIELD_VALUE_URL
   //     | BOOLEAN
   //     | DIGIT
@@ -1599,6 +1600,7 @@ public class ImpexParser implements PsiParser, LightPsiParser {
     if (!r) r = consumeToken(b, FIELD_VALUE_JAR_PREFIX);
     if (!r) r = consumeToken(b, FIELD_VALUE_FILE_PREFIX);
     if (!r) r = consumeToken(b, FIELD_VALUE_ZIP_PREFIX);
+    if (!r) r = consumeToken(b, FIELD_VALUE_HTTP_PREFIX);
     if (!r) r = consumeToken(b, FIELD_VALUE_URL);
     if (!r) r = consumeToken(b, BOOLEAN);
     if (!r) r = consumeToken(b, DIGIT);