+++ gen/com/intellij/idea/plugin/hybris/impex/psi/ImpexTypes.java
@@ -90,6 +90,7 @@ public interface ImpexTypes {
   IElementType FIELD_LIST_ITEM_SEPARATOR = new ImpexTokenType("FIELD_LIST_ITEM_SEPARATOR");
   IElementType FIELD_VALUE = new ImpexTokenType("FIELD_VALUE");
   IElementType FIELD_VALUE_FILE_PREFIX = new ImpexTokenType("FIELD_VALUE_FILE_PREFIX");
+  IElementType FIELD_VALUE_HTTP_PREFIX = new ImpexTokenType("FIELD_VALUE_HTTP_PREFIX");
   IElementType FIELD_VALUE_IGNORE = new ImpexTokenType("FIELD_VALUE_IGNORE");
   IElementType FIELD_VALUE_JAR_PREFIX = new ImpexTokenType("FIELD_VALUE_JAR_PREFIX");
   IElementType FIELD_VALUE_NULL = new ImpexTokenType("FIELD_VALUE_NULL");